+++ Mirle.Agv/Model/Map/MapSection.cs
@@ -13,8 +13,8 @@ namespace Mirle.Agv.Model
         public string Id { get; set; } = "Empty";
         public MapAddress HeadAddress { get; set; } = new MapAddress();
         public MapAddress TailAddress { get; set; } = new MapAddress();
-        public float Distance { get; set; }
-        public float Speed { get; set; }
+        public double Distance { get; set; }
+        public double Speed { get; set; }
         public EnumSectionType Type { get; set; } = EnumSectionType.None;
         public EnumPermitDirection PermitDirection { get; set; } = EnumPermitDirection.None;
         public bool FowardBeamSensorDisable { get; set; }
