+++ Mirle.Agv/Model/MoveModel/Sr2000/BarcodeData.cs
@@ -18,7 +18,7 @@ namespace Mirle.Agv.Model
         public BarcodeData(int id, double x, double y)
         {
             ID = id;
-            ViewPosition = new MapPosition((float)x, (float)y);
+            ViewPosition = new MapPosition(x, y);
         }
     }
 }
